remove showtable, refactor fetching logic for recently played
